+++ Classes/CalcBreakdownControl.lua
@@ -351,8 +351,6 @@ function CalcBreakdownClass:AddModSection(sectionData, modList)
 				local node = build.spec.nodes[tonumber(nodeId)]
 				row.sourceName = node.dn
 				row.sourceNameNode = node
-			elseif row.mod.source == "Tree:Jewel" then
-				row.sourceName = "Jewel conversion"
 			end
 		elseif sourceType == "Skill" then
 			-- Extract skill name
@@ -389,6 +387,8 @@ function CalcBreakdownClass:AddModSection(sectionData, modList)
 					else
 						desc = baseVal.." per "..self:FormatModName(tag.var)
 					end
+				elseif tag.type == "MultiplierThreshold" then
+					desc = "If "..self:FormatModName(tag.var)..(tag.upper and " <= " or " >= ")..(tag.threshold or self:FormatModName(tag.thresholdVar))
 				elseif tag.type == "PerStat" then
 					if tag.base then
 						desc = (row.mod.type == "BASE" and string.format("%+g", tag.base) or tag.base.."%").." + "..math.abs(row.mod.value).." per "..(tag.div or 1).." "..self:FormatModName(tag.var)
@@ -396,7 +396,7 @@ function CalcBreakdownClass:AddModSection(sectionData, modList)
 						desc = baseVal.." per "..(tag.div or 1).." "..self:FormatModName(tag.stat)
 					end
 				elseif tag.type == "StatThreshold" then
-					desc = "If "..self:FormatModName(tag.stat).." >= "..tag.threshold
+					desc = "If "..self:FormatModName(tag.stat)..(tag.upper and " <= " or " >= ")..(tag.threshold or self:FormatModName(tag.thresholdStat))
 				elseif tag.type == "SkillName" then
 					desc = "Skill: "..tag.skillName
 				elseif tag.type == "SkillId" then